refactor(planning): extract the desktop planning row components (#467)
Desktop counterpart to the mobile rows split. Move DesktopPlanningView's table row
components — DGoalItemRow (goal item, kebab menu), DPlanRow (the generic
fixed/insurance row), and the shared kebab MenuBtn — into a new
`desktopPlanningRows.tsx`. Also drop the desktop copy of EditIcon / TrashIcon and
reuse the shared `planningIcons.tsx` (introduced with the mobile split). The
components are module-level and prop-driven, so this is a relocation — the view
still owns the data + handlers. No behavior or markup change; now-unused lucide
imports (MoreHorizontal / Check / TrendingUp) and useCloseOnScroll drop out.

Parity: the PlanningClient / DesktopPlanningView component tests + the
planning-desktop E2E (DCA fund grouped under its goal — renders DGoalItemRow).

Verified: full unit suite 1261 passed; typecheck 0 errors; lint 0 errors;
git diff --check clean; E2E planning-desktop (2) passed. DesktopPlanningView
1135 → 936 lines.
